Major update: cylindrical algae
BioSNICAR_GO now has functionality for including cylindricla algae of any dimensions. The radius and length are user defined in the BiOSNICAR_GO driver and used to retrieve the optical properties from a library of NetCDF files that is populated using the python script Algae_GO.py. The optical properties have been determine dusing the geometric optics approximation for cylinders.